implement repo route handle

+++ src/domain/repo_info/mod.rs
@@ -9,7 +9,7 @@ pub struct RepoInfo {
     pub branch: String,
     pub status: String,
     pub last_commit: String,
-    pub timeout: i64,
+    pub timeout: i32,
     pub modified_time: i64,
 }
